debug client done
