+++ modules/tools/map_datachecker/client.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 ###############################################################################
-# Copyright 2017 The Apollo Authors. All Rights Reserved.
+# Copyright 2019 The Apollo Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,29 +19,6 @@
 set -ue
 SCRIPT_PATH="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 APOLLO_ROOT_PATH="${SCRIPT_PATH}/../../.."
-MAP_DATACHECKER=${APOLLO_ROOT_PATH}/bazel-bin/modules/map/tools/map_datachecker/map_datachecker
-CONF=${APOLLO_ROOT_PATH}/modules/map/tools/map_datachecker/conf/map-datachecker.conf
-${MAP_DATACHECKER} --flagfile=${CONF} > ${SCRIPT_PATH}/`date '+%Y-%m-%d--%H-%M-%S'`.log 2>&1 &
-echo 'Server has been started successfully'
-echo 'You can enter Ctrl+C to exit this server'
-PID=$!
-trap "trap_func" INT
-function trap_func() {
-  echo 'Stopping map_datachecker......'
-  kill -INT ${PID}
-}
-while [[ true ]]; do
-  sleep 2
-done
-
-
-
-
-
-
-
-
-
-
-
-
+LOG_DIR=${APOLLO_ROOT_PATH}/modules/tools/map_datachecker/log
+PYTHON_SCRIPT_PATH=${SCRIPT_PATH}/python
+python ${PYTHON_SCRIPT_PATH}/client.py $@ >> ${LOG_DIR}/client.log
