+++ modules/tools/map_datachecker/conf/state_machine.yaml
