+++ modules/tools/map_datachecker/python/channel_check.py
@@ -26,11 +26,11 @@ import multiprocessing
 import yaml
 import time
 script_path = os.path.dirname(os.path.realpath(__file__))
-apollo_root = os.path.join(script_path, "../../..")
+apollo_root = os.path.join(script_path, "../../../..")
 pb_path = os.path.join(apollo_root, "py_proto/modules/map/tools/map_datachecker/proto/")
 sys.path.append(pb_path)
 import collection_service_pb2_grpc
-import collection_service_pb2
+# import collection_service_pb2
 import collection_check_message_pb2
 import collection_error_code_pb2 as ErrorCode
 
@@ -41,8 +41,9 @@ class ChannelChecker:
             self._conf = yaml.load(fp.read())["channel_check"]
             self._host = self._conf['grpc_host']
             self._port = self._conf['grpc_port']
-        channel = grpc.insecure_channel(self._host + ':' + self._port)
-        self.stub = collection_service_pb2_grpc.CollectionCheckerServiceStub(channel)
+        logging.info("ChannelChecker grpc socket: " + self._host + ':' + self._port)
+        self.channel = grpc.insecure_channel(self._host + ':' + self._port)
+        self.stub = collection_service_pb2_grpc.CollectionCheckerServiceStub(self.channel)
 
     def _exception_handler(self, error_code):
         if error_code == ErrorCode.SUCCESS:
@@ -101,8 +102,6 @@ class ChannelChecker:
             logging.info("channel checker sleep %d seconds" % self._conf["check_period"])
             time.sleep(self._conf["check_period"])
         logging.info("detected stop flag file, periodically checking will exit")
-        
-
 
     def async_start(self, record_path):
         if not record_path or not os.path.exists(record_path):
@@ -125,16 +124,17 @@ class ChannelChecker:
         if ret != 0:
             logging.error("start check channel failed, record_path [%s]" % record_path)
             return -1
+        self._periodic_check()
         # thread = threading.Thread(target=self._periodic_check)
         # thread.start()
-        process = multiprocessing.Process(target=self._periodic_check)
-        process.start()
-        logging.info("async start checking thread, main thread will exit")
+        # process = multiprocessing.Process(target=self._periodic_check)
+        # process.start()
+        # logging.info("async start checking thread, main thread will exit")
 
-#        import state_machine
-#        state_machine.StateMachine()
-#        state_checker = StateMachine()
-#        state_checker.state_flow('record_check_start', 'start', 'record_check_start')
+        # import state_machine
+        # state_machine.StateMachine()
+        # state_checker = StateMachine()
+        # state_checker.state_flow('record_check_start', 'start', 'record_check_start')
         return 0
 
     def async_stop(self):
